Enhancement: Move ErrorsManager into Symfony\CS\Error namespace
